+++ tests/ILSpyMcp.Tests/ILSpyMcpCmdTests.cs
@@ -25,6 +25,7 @@ public class ILSpyMcpCmdTests
         var token = TestDataPaths.FirstCalleeMethodToken(TestDataPaths.TestSamplesDll);
         var result = await ILSpyMcpCmd.DispatchCliAsync(
             assembly: TestDataPaths.TestSamplesDll, typeName: "", memberName: "", entityTypes: "", nameContains: "", namespaceContains: "",
+            searchString: "",
             outputDir: "", project: false, nestedDirectories: false, signatures: false, hierarchy: false,
             dependencies: false, callGraph: true, external: false, indirect: false, assemblyInfo: false,
             token: token, typeToken: "", lines: "", timeoutSeconds: 30, check: false);